Remove option-less producer as it seems to be useless
