+++ src/Core/src/Eventuous.Producers/Diagnostics/ProducerEventSource.cs
@@ -9,7 +9,7 @@ using Eventuous.Diagnostics;
 namespace Eventuous.Producers.Diagnostics;
 
 [EventSource(Name = $"{DiagnosticName.BaseName}.producer")]
-public class ProducerEventSource<T> : EventSource where T : IEventProducer {
+public class ProducerEventSource<T> : EventSource where T : class {
     public static readonly ProducerEventSource<T> Log = new();
     
     static readonly string ProducerName = typeof(T).Name;
