+++ src/Gateway/src/Eventuous.Gateway/Eventuous.Gateway.csproj
@@ -3,20 +3,20 @@
         <GenerateDocumentationFile>true</GenerateDocumentationFile>
     </PropertyGroup>
     <ItemGroup>
-        <ProjectReference Include="$(CoreRoot)\Eventuous.Producers\Eventuous.Producers.csproj"/>
-        <ProjectReference Include="$(CoreRoot)\Eventuous.Subscriptions\Eventuous.Subscriptions.csproj"/>
+        <ProjectReference Include="$(CoreRoot)\Eventuous.Producers\Eventuous.Producers.csproj" />
+        <ProjectReference Include="$(CoreRoot)\Eventuous.Subscriptions\Eventuous.Subscriptions.csproj" />
     </ItemGroup>
     <ItemGroup>
-        <Using Include="Eventuous.Producers"/>
-        <Using Include="Eventuous.Subscriptions"/>
+        <Using Include="Eventuous.Producers" />
+        <Using Include="Eventuous.Subscriptions" />
     </ItemGroup>
     <ItemGroup>
-        <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions"/>
+        <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" />
     </ItemGroup>
     <ItemGroup>
         <Compile Include="$(CoreRoot)\Eventuous.Shared\Tools\TaskExtensions.cs">
             <Link>Tools\TaskExtensions.cs</Link>
         </Compile>
-        <Using Include="Eventuous.Tools"/>
+        <Using Include="Eventuous.Tools" />
     </ItemGroup>
 </Project>
