+++ src/Gateway/src/Eventuous.Gateway/GatewayProducer.cs
@@ -3,11 +3,14 @@
 
 namespace Eventuous.Gateway;
 
-class GatewayProducer<T> : GatewayProducer, IEventProducer<T> where T : class {
+class GatewayProducer<T> : IEventProducer<T> where T : class {
     readonly IEventProducer<T> _inner;
+    readonly bool              _isHostedService;
 
-    public GatewayProducer(IEventProducer<T> inner) : base(inner)
-        => _inner = inner;
+    public GatewayProducer(IEventProducer<T> inner) {
+        _isHostedService = inner is not IHostedProducer;
+        _inner           = inner;
+    }
 
     public async Task Produce(StreamName stream, IEnumerable<ProducedMessage> messages, T? options, CancellationToken cancellationToken = default) {
         if (_isHostedService) {
@@ -16,27 +19,8 @@ class GatewayProducer<T> : GatewayProducer, IEventProducer<T> where T : class {
 
         await _inner.Produce(stream, messages, options, cancellationToken).NoContext();
     }
-}
 
-class GatewayProducer : IEventProducer {
-    readonly IEventProducer _inner;
-
-    protected readonly bool _isHostedService;
-
-    public GatewayProducer(IEventProducer inner) {
-        _isHostedService = inner is not IHostedProducer;
-        _inner           = inner;
-    }
-
-    public async Task Produce(StreamName stream, IEnumerable<ProducedMessage> messages, CancellationToken cancellationToken = default) {
-        if (_isHostedService) {
-            await WaitForInner(_inner, cancellationToken).NoContext();
-        }
-
-        await _inner.Produce(stream, messages, cancellationToken).NoContext();
-    }
-
-    protected static async ValueTask WaitForInner(IEventProducer inner, CancellationToken cancellationToken) {
+    static async ValueTask WaitForInner(IEventProducer<T> inner, CancellationToken cancellationToken) {
         if (inner is not IHostedProducer hosted) return;
 
         while (!hosted.Ready) {
@@ -44,4 +28,7 @@ class GatewayProducer : IEventProducer {
             await Task.Delay(1000, cancellationToken).NoContext();
         }
     }
+
+    public Task Produce(StreamName stream, IEnumerable<ProducedMessage> messages, CancellationToken cancellationToken = default)
+        => Produce(stream, messages, null, cancellationToken);
 }
